Implement UI of Collect Fine, but there's still bugs in operation

+++ LibraryManagement/Models/Book.cs
@@ -21,12 +21,12 @@ namespace LibraryManagement.Models
             this.DetailBillReturns = new HashSet<DetailBillReturn>();
             this.Authors = new HashSet<Author>();
         }
-        
+    
         public int idBook { get; set; }
-        public string nameBook { get ; set; }
+        public string nameBook { get; set; }
         public System.DateTime dateManufacture { get; set; }
         public System.DateTime dateAddBook { get; set; }
-        public double price { get; set; }
+        public int price { get; set; }
         public string statusBook { get; set; }
         public int idCategory { get; set; }
         public int idPublisher { get; set; }
